Feature added: emulate open(dir) for nuttx and tizenrt. There is a range of fd values that such a descriptor can have.
The returned value is not seen as valid in the underlying OS, it only
works within libtuv.

Not thread-safe for now (TODO).
Due to the implementation (storing of paths) this is vulnerable to
renames between open and actual action (fstat, [F]utime).

libtuv-DCO-1.0-Signed-off-by: Tomasz Wozniak t.wozniak@samsung.com